requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
decorator==4.4.2
Django==3.0.9
django-admin-easy==0.6
django-admin-rangefilter==0.8.1
django-audiofield==0.10.0
django-celery==3.3.1
django-filter==2.3.0
django-rest-framework==0.1.0
djangorestframework==3.11.1
drf-yasg==1.17.1
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
mutagen==1.45.1
openapi-codec==1.3.2
packaging==20.4
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pycparser==2.20
pyparsing==2.4.7
python-telegram-bot==12.6.1
pytz==2020.1
reportlab==3.5.47
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
tornado==6.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
